django_app/requirements/tests.txt
coverage==5.5
django-coverage-plugin==1.8.0
django-test-plus==1.4.0
xeger
factory_boy==3.2.0
pytest-django==4.2.0
pytest-sugar==0.9.4
pytest-cov==2.11.1
pytest-env==0.6.2
fakeredis==1.5.0
